requirements-test.txt
responses
djangorestframework-simplejwt
django-cors-headers>=2.5.2
packaging
immunity-sphinx-theme~=1.0.0
freezegun
django-extensions
immunity-utils[qa] @ https://github.com/edge-servers/immunity-utils/tarball/master
pylinkvalidator
lxml~=4.9.1
cssselect~=1.1.0
